# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Koupreng E-Invitation API
  description: |
    REST API for the Koupreng E-Invitation platform, including authentication,
    users, invitations, templates, guests, RSVP, check-in, media, payments,
    organizations, subscriptions, notifications, reporting, administration,
    and related services. The current contract includes both legacy /api routes
    and versioned /api/v1 routes; the version below does not imply a global base path.
  version: v1
servers:
- url: http://localhost
  description: Generated server url
...
paths:
...
  /api/v1/public/invitations/{slug}/rsvp:
    post:
      tags:
      - RSVP
      summary: Submit a public RSVP
      description: Create or update a public RSVP after invitation access and deadline
        validation.
      operationId: publicRsvp
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        example: demo-wedding
      - name: accessToken
        in: query
        required: false
        schema:
          type: string
        example: demo-invitation-access-token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpRequest"
            examples:
              attending:
                summary: Attending with a guest
                value:
                  guestName: Dara Sok
                  phone: "012345679"
                  email: dara.sok@example.com
                  responseStatus: ATTENDING
                  attendeeCount: 2
                  message: We are delighted to celebrate with you.
              notAttending:
                summary: Unable to attend
                value:
                  guestName: Sophea Lim
                  phone: "012345682"
                  email: sophea.lim@example.com
                  responseStatus: NOT_ATTENDING
                  attendeeCount: 0
                  message: Thank you for the invitation; we send our best wishes.
        required: true
      responses:
        "201":
          description: RSVP accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRsvpResponse"
        "400":
          description: Invalid RSVP or closed deadline
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Public RSVP rate limit exceeded
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiError"
      security: []
  /api/v1/public/invitations/{slug}/guests/{inviteToken}/rsvp:
    post:
      tags:
      - RSVP
      summary: Submit a personalized guest RSVP
      description: Create or update the RSVP associated with an opaque guest invitation
        token.
      operationId: publicTokenRsvp
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        example: demo-wedding
      - name: inviteToken
        in: path
        required: true
        schema:
          type: string
        example: demo-guest-attending-token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpRequest"
            examples:
              attending:
                summary: Attending with a guest
                value:
                  guestName: Dara Sok
                  phone: "012345679"
                  email: dara.sok@example.com
                  responseStatus: ATTENDING
                  attendeeCount: 2
                  message: We are delighted to celebrate with you.
              notAttending:
                summary: Unable to attend
                value:
                  guestName: Sophea Lim
                  phone: "012345682"
                  email: sophea.lim@example.com
                  responseStatus: NOT_ATTENDING
                  attendeeCount: 0
                  message: Thank you for the invitation; we send our best wishes.
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRsvpResponse"
      security: []
...
  /api/v1/invitations/{invitationId}/rsvps/{rsvpId}:
    delete:
      tags:
      - RSVP
      summary: Delete a rsvp record
      operationId: delete_3
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: rsvpId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
      security:
      - bearerAuth: []
    patch:
      tags:
      - RSVP
      summary: Update a rsvp record
      operationId: update_3
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: rsvpId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpUpdateRequest"
            examples:
              attending:
                summary: Mark the RSVP as attending
                value:
                  responseStatus: ATTENDING
                  attendeeCount: 2
                  message: We look forward to celebrating with you.
              notAttending:
                summary: Mark the RSVP as not attending
                value:
                  responseStatus: NOT_ATTENDING
                  attendeeCount: 0
                  message: Thank you for inviting us.
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRsvpResponse"
      security:
      - bearerAuth: []
...
  /api/v1/public/invitations/{slug}/wishes:
    get:
      tags:
      - RSVP
      summary: List public guest wishes
      description: Return approved wishes for an accessible published invitation.
      operationId: publicWishes
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        example: demo-wedding
      - name: accessToken
        in: query
        required: false
        schema:
          type: string
        example: demo-invitation-access-token
      - name: token
        in: query
        required: false
        schema:
          type: string
        example: demo-guest-attending-token
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListWishResponse"
      security: []
  /api/v1/public/invitations/{slug}/rsvp-summary-public:
    get:
      tags:
      - RSVP
      summary: Get the public RSVP summary
      operationId: publicSummary
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        example: demo-wedding
      - name: accessToken
        in: query
        required: false
        schema:
          type: string
        example: demo-invitation-access-token
      - name: token
        in: query
        required: false
        schema:
          type: string
        example: demo-guest-attending-token
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRsvpSummaryResponse"
      security: []
...
  /api/v1/notifications/summary:
    get:
      tags:
      - Notifications
      summary: Summarize notification activity
      operationId: summary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseNotificationSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/wishes:
    get:
      tags:
      - RSVP
      summary: List invitation wishes
      operationId: wishes
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListWishResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/seating/summary:
    get:
      tags:
      - Seating
      summary: Summarize seating activity
      operationId: summary_1
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeatingSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/rsvps:
    get:
      tags:
      - RSVP
      summary: List rsvp records
      operationId: list_3
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListRsvpResponse"
      security:
      - bearerAuth: []
  /api/v1/invitations/{invitationId}/rsvps/summary:
    get:
      tags:
      - RSVP
      summary: Summarize rsvp activity
      operationId: summary_2
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRsvpSummaryResponse"
      security:
      - bearerAuth: []
  /api/v1/invitations/{invitationId}/reports/rsvp:
    get:
      tags:
      - Reports
      summary: Rsvp Report
      operationId: rsvpReport
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRsvpReportResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/check-in/summary:
    get:
      tags:
      - Check-In
      summary: Summarize check in activity
      operationId: summary_4
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCheckInSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/admin/reports/rsvp:
    get:
      tags:
      - Administration
      summary: Rsvp Report
      operationId: rsvpReport_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAdminReportResponse"
      security:
      - bearerAuth: []
...
  /api/v1/admin/analytics/rsvp:
    get:
      tags:
      - Administration
      summary: Analytics Rsvp
      operationId: analyticsRsvp
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAdminReportResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/wishes/{wishId}:
    delete:
      tags:
      - RSVP
      summary: Remove wish text while preserving the RSVP attendance record
      operationId: moderateWish
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: wishId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MessageResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    RsvpRequest:
      type: object
      properties:
        guestName:
          type: string
        phone:
          type: string
        email:
          type: string
        responseStatus:
          type: string
          enum:
          - ATTENDING
          - NOT_ATTENDING
          - MAYBE
        attendeeCount:
          type: integer
          format: int32
          minimum: 0
        message:
          type: string
      required:
      - responseStatus
    ApiError:
      type: object
      description: Standard error body returned by the API exception handler.
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
          example: 400
        error:
          type: string
          example: Bad Request
        code:
          type: string
          example: REQUEST_INVALID
        message:
          type: string
          example: Request validation failed
        path:
          type: string
          example: /api/v1/invitations
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        fields:
          type: object
          additionalProperties:
            type: string
          description: "Legacy validation-field alias, present only on validation\
            \ errors."
    ApiResponseRsvpResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/RsvpResponse"
        timestamp:
          type: string
          format: date-time
    RsvpResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        invitationId:
          type: integer
          format: int64
        guestId:
          type: integer
          format: int64
        guestName:
          type: string
        inviteToken:
          type: string
        qrCodeUrl:
          type: string
        responseStatus:
          type: string
          enum:
          - ATTENDING
          - NOT_ATTENDING
          - MAYBE
        attendeeCount:
          type: integer
          format: int32
        message:
          type: string
        respondedAt:
          type: string
          format: date-time
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    RsvpUpdateRequest:
      type: object
      properties:
        responseStatus:
          type: string
          enum:
          - ATTENDING
          - NOT_ATTENDING
          - MAYBE
        attendeeCount:
          type: integer
          format: int32
          minimum: 0
        message:
          type: string
...
    ApiResponseListWishResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/WishResponse"
        timestamp:
          type: string
          format: date-time
    WishResponse:
      type: object
      properties:
        rsvpId:
          type: integer
          format: int64
        guestId:
          type: integer
          format: int64
        guestName:
          type: string
        message:
          type: string
        respondedAt:
          type: string
          format: date-time
    ApiResponseRsvpSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/RsvpSummaryResponse"
        timestamp:
          type: string
          format: date-time
    RsvpSummaryResponse:
      type: object
      properties:
        totalGuests:
          type: integer
          format: int64
        attending:
          type: integer
          format: int64
        notAttending:
          type: integer
          format: int64
        maybe:
          type: integer
          format: int64
        pending:
          type: integer
          format: int64
        totalAttendeeCount:
          type: integer
          format: int64
...
    ApiResponseNotificationSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/NotificationSummaryResponse"
        timestamp:
          type: string
          format: date-time
    NotificationSummaryResponse:
      type: object
      properties:
        total:
          type: integer
          format: int64
        unread:
          type: integer
          format: int64
        pending:
          type: integer
          format: int64
        sent:
          type: integer
          format: int64
        delivered:
          type: integer
          format: int64
        failed:
          type: integer
          format: int64
        cancelled:
          type: integer
          format: int64
...
    ApiResponseSeatingSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/SeatingSummaryResponse"
        timestamp:
          type: string
          format: date-time
    SeatingSummaryResponse:
      type: object
      properties:
        invitationId:
          type: integer
          format: int64
        totalTables:
          type: integer
          format: int64
        totalCapacity:
          type: integer
          format: int64
        assignedSeats:
          type: integer
          format: int64
        remainingSeats:
          type: integer
          format: int64
    ApiResponseListRsvpResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/RsvpResponse"
        timestamp:
          type: string
          format: date-time
    ApiResponseRsvpReportResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/RsvpReportResponse"
        timestamp:
          type: string
          format: date-time
    RsvpReportResponse:
      type: object
      properties:
        invitationId:
          type: integer
          format: int64
        totalGuests:
          type: integer
          format: int64
        yesCount:
          type: integer
          format: int64
        noCount:
          type: integer
          format: int64
        maybeCount:
          type: integer
          format: int64
        pendingCount:
          type: integer
          format: int64
        attendeeTotal:
          type: integer
          format: int64
        responses:
          type: array
          items:
            $ref: "#/components/schemas/RsvpResponse"
...
    ApiResponseCheckInSummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/CheckInSummaryResponse"
        timestamp:
          type: string
          format: date-time
    CheckInSummaryResponse:
      type: object
      properties:
        invitationId:
          type: integer
          format: int64
        totalGuests:
          type: integer
          format: int64
        checkedIn:
          type: integer
          format: int64
        remaining:
          type: integer
          format: int64
        attendingCheckedIn:
          type: integer
          format: int64
...
    AdminReportResponse:
      type: object
      properties:
        report:
          type: string
        generatedAt:
          type: string
          format: date-time
        summary:
          type: object
          additionalProperties: {}
        rows:
          type: array
          items: {}
    ApiResponseAdminReportResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/AdminReportResponse"
        timestamp:
          type: string
          format: date-time
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      description: Paste the accessToken returned by POST /api/auth/login. Scalar
        adds the Bearer prefix automatically; do not type it twice.
      scheme: bearer
      bearerFormat: JWT
    adminPaymentSecret:
      type: apiKey
      description: Internal Telegram payment reconciliation secret.
      name: X-ADMIN-PAYMENT-SECRET
      in: header